AI_Employee_Vault/bronze/pyproject.toml
[project]
name = "ai-employee-vault"
version = "0.1.0"
description = "Bronze Tier AI Employee - File Processing Automation"
requires-python = ">=3.13"
dependencies = [
    "watchdog>=4.0.0",
    "pyyaml>=6.0.1",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
